+++ aquadoggo/src/network/config.rs
@@ -5,6 +5,7 @@ use std::path::PathBuf;
 use anyhow::Result;
 use libp2p::identity::Keypair;
 use libp2p::swarm::ConnectionLimits;
+use libp2p::{Multiaddr, PeerId};
 use log::info;
 use serde::{Deserialize, Serialize};
 
@@ -12,6 +13,9 @@ use crate::network::identity::Identity;
 
 /// Key pair file name.
 const KEY_PAIR_FILE_NAME: &str = "private-key";
+
+/// The namespace used by the `identify` network behaviour.
+pub const NODE_NAMESPACE: &str = "aquadoggo";
 
 /// Network config for the node.
 #[derive(Debug, Clone, Deserialize, Serialize)]
@@ -71,7 +75,23 @@ pub struct NetworkConfiguration {
     pub quic_port: u16,
 
     /// The addresses of remote peers to replicate from.
-    pub remote_peers: Vec<String>,
+    pub remote_peers: Vec<Multiaddr>,
+
+    /// Rendezvous client behaviour enabled.
+    ///
+    /// Connect to a rendezvous point, register the local node and query addresses of remote peers.
+    pub rendezvous_client: bool,
+
+    /// Rendezvous server behaviour enabled.
+    ///
+    /// Serve as a rendezvous point for peer discovery, allowing peer registration and queries.
+    pub rendezvous_server: bool,
+
+    /// Address of a rendezvous server in the form of a multiaddress.
+    pub rendezvous_address: Option<Multiaddr>,
+
+    /// Peer ID of a rendezvous server.
+    pub rendezvous_peer_id: Option<PeerId>,
 }
 
 impl Default for NetworkConfiguration {
@@ -89,6 +109,10 @@ impl Default for NetworkConfiguration {
             ping: false,
             quic_port: 2022,
             remote_peers: Vec::new(),
+            rendezvous_client: false,
+            rendezvous_server: false,
+            rendezvous_address: None,
+            rendezvous_peer_id: None,
         }
     }
 }